+++ src/client/camera.h
@@ -118,19 +118,19 @@ public:
 	}
 
 	// Horizontal field of view
-	inline f32 getFovX() const
+	inline f64 getFovX() const
 	{
 		return m_fov_x;
 	}
 
 	// Vertical field of view
-	inline f32 getFovY() const
+	inline f64 getFovY() const
 	{
 		return m_fov_y;
 	}
 
 	// Get maximum of getFovX() and getFovY()
-	inline f32 getFovMax() const
+	inline f64 getFovMax() const
 	{
 		return MYMAX(m_fov_x, m_fov_y);
 	}
@@ -142,7 +142,7 @@ public:
 	{
 		return [planes = getFrustumCullPlanes(),
 				camera_offset = intToFloat(m_camera_offset, BS)
-				](v3d position, f32 radius) {
+				](v3d position, f64 radius) {
 			v3d pos_camspace = position - camera_offset;
 			for (auto &plane : planes) {
 				if (plane.getDistanceTo(pos_camspace) > radius)
@@ -156,10 +156,10 @@ public:
 	void notifyFovChange();
 
 	// Step the camera: updates the viewing range and view bobbing.
-	void step(f32 dtime);
+	void step(f64 dtime);
 
 	// Update the camera from the local player's position.
-	void update(LocalPlayer* player, f32 frametime, f32 tool_reload_ratio);
+	void update(LocalPlayer* player, f64 frametime, f64 tool_reload_ratio);
 
 	// Update render distance
 	void updateViewingRange();
@@ -206,12 +206,12 @@ public:
 
 	void drawNametags();
 
-	inline void addArmInertia(f32 player_yaw);
+	inline void addArmInertia(f64 player_yaw);
 
 private:
 	// Use getFrustumCuller().
 	// This helper just exists to decrease the header's number of includes.
-	std::array<core::plane3d<f32>, 4> getFrustumCullPlanes() const;
+	std::array<core::plane3d<f64>, 4> getFrustumCullPlanes() const;
 
 	// Nodes
 	scene::ISceneNode *m_playernode = nullptr;
@@ -227,7 +227,7 @@ private:
 	Client *m_client;
 
 	// Default Client FOV (as defined by the "fov" setting)
-	f32 m_cache_fov;
+	f64 m_cache_fov;
 
 	// Absolute camera position
 	v3d m_camera_position;
@@ -240,11 +240,11 @@ private:
 
 	// Server-sent FOV variables
 	bool m_server_sent_fov = false;
-	f32 m_curr_fov_degrees, m_old_fov_degrees, m_target_fov_degrees;
+	f64 m_curr_fov_degrees, m_old_fov_degrees, m_target_fov_degrees;
 
 	// FOV transition variables
 	bool m_fov_transition_active = false;
-	f32 m_fov_diff, m_transition_time;
+	f64 m_fov_diff, m_transition_time;
 
 	v2f m_wieldmesh_offset = v2f(55.0f, -35.0f);
 	v2f m_arm_dir;
@@ -253,36 +253,36 @@ private:
 	v2f m_last_cam_pos;
 
 	// Field of view and aspect ratio stuff
-	f32 m_aspect = 1.0f;
-	f32 m_fov_x = 1.0f;
-	f32 m_fov_y = 1.0f;
+	f64 m_aspect = 1.0f;
+	f64 m_fov_x = 1.0f;
+	f64 m_fov_y = 1.0f;
 
 	// View bobbing animation frame (0 <= m_view_bobbing_anim < 1)
-	f32 m_view_bobbing_anim = 0.0f;
+	f64 m_view_bobbing_anim = 0.0f;
 	// If 0, view bobbing is off (e.g. player is standing).
 	// If 1, view bobbing is on (player is walking).
 	// If 2, view bobbing is getting switched off.
 	s32 m_view_bobbing_state = 0;
 	// Speed of view bobbing animation
-	f32 m_view_bobbing_speed = 0.0f;
+	f64 m_view_bobbing_speed = 0.0f;
 	// Fall view bobbing
-	f32 m_view_bobbing_fall = 0.0f;
+	f64 m_view_bobbing_fall = 0.0f;
 
 	// Digging animation frame (0 <= m_digging_anim < 1)
-	f32 m_digging_anim = 0.0f;
+	f64 m_digging_anim = 0.0f;
 	// If -1, no digging animation
 	// If 0, left-click digging animation
 	// If 1, right-click digging animation
 	s32 m_digging_button = -1;
 
 	// Animation when changing wielded item
-	f32 m_wield_change_timer = 0.125f;
+	f64 m_wield_change_timer = 0.125f;
 	ItemStack m_wield_item_next;
 
 	CameraMode m_camera_mode = CAMERA_MODE_FIRST;
 
-	f32 m_cache_fall_bobbing_amount;
-	f32 m_cache_view_bobbing_amount;
+	f64 m_cache_fall_bobbing_amount;
+	f64 m_cache_view_bobbing_amount;
 	bool m_arm_inertia;
 
 	std::list<Nametag *> m_nametags;
